delivery option

+++ database/migrations/2019_08_15_120216_create_payments_table.php
@@ -19,7 +19,7 @@ class CreatePaymentsTable extends Migration
             $table->unsignedBigInteger('order_id');
             $table->string('payment_link')->nullable()->default(null);
             $table->dateTime('paid_at')->nullable()->default(null);
-            $table->decimal('amount');
+            $table->float('amount');
             $table->timestamps();
 
             $table->foreign('merchant_id')->references('id')->on('merchants');
